pyproject.toml

[project]
name = ""
version = ""
description = ""
authors = [
    {name = "Sinkerine", email = "git@15cm.net"},
]
dependencies = [
    "django>=4.2.2",
    "django-ninja>=0.22.2",
    "psycopg[binary]>=3.1.9",
    "django-annoying>=0.10.6",
    "celery[redis]>=5.3.1",
]
requires-python = ">=3.10"
license = {text = "MIT"}

[tool.pdm.dev-dependencies]
dev = [
    "django-stubs>=4.2.3",
    "django-extensions>=3.2.3",
    "daphne>=4.0.0",
]
test = [
    "pytest>=7.4.0",
    "pytest-dotenv>=0.5.2",
    "pytest-django>=4.5.2",
    "decorator>=5.1.1",
]
